PCB cleanup silkscreen, fab-layer
on the fabrication layer all the values of the small components like
resistors and capactiors are now inside the footprints for easier
referencing.

the silk screen markings have been cleaned up as well, most notably
moved them off the footprints on hardpass, and aligned them neatly on
hardpass-sci, where some of the automatically placed ones were even on
top of solder pads.